size moved to wrapper

+++ src/hotspot/share/classfile/javaClasses.hpp
@@ -181,7 +181,7 @@ class java_lang_String : AllStatic {
     return h;
   }
 
-  static unsigned int hash_code(const char *utf8_str, int utf8_len) {
+  static unsigned int hash_code(const char *utf8_str, size_t utf8_len) {
     unsigned int h = 0;
     int unicode_length = UTF8::unicode_length(utf8_str, utf8_len);
 
@@ -198,7 +198,7 @@ class java_lang_String : AllStatic {
 
   // Compare strings (of different types/encodings), length is the string (array) length
   static bool equals(oop java_string, const jchar* chars, int len);
-  static bool equals(oop java_string, const char* utf8_str, int utf8_len);
+  static bool equals(oop java_string, const char* utf8_str, size_t utf8_len);
   static bool equals(oop str1, oop str2);
   static inline bool value_equals(typeArrayOop str_value1, typeArrayOop str_value2);
